requirements.txt
aiogram==3.4.1
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
aiohttp==3.9.3
python-dotenv==1.0.0
alembic==1.13.1
pydantic>=2.4.1,<2.6
pydantic-settings>=2.0.0,<2.2
google-genai>=1.0.0
